Extract participations section
